+++ code/src/sixsq/nuvla/server/resources/group.clj
@@ -26,8 +26,9 @@ that start with 'nuvla-' are reserved for the server.
 (def ^:const create-type (u/ns->create-type *ns*))
 
 
-(def collection-acl {:query ["group/nuvla-user"]
-                     :add   ["group/nuvla-admin"]})
+(def collection-acl {:query       ["group/nuvla-user"]
+                     :add         ["group/nuvla-admin"]
+                     :bulk-delete ["group/nuvla-admin"]})
 
 
 ;;
@@ -137,6 +138,14 @@ that start with 'nuvla-' are reserved for the server.
 (defmethod crud/query resource-type
   [request]
   (query-impl request))
+
+
+(def bulk-delete-impl (std-crud/bulk-delete-fn resource-type collection-acl collection-type))
+
+
+(defmethod crud/bulk-delete resource-type
+  [request]
+  (bulk-delete-impl request))
 
 
 ;;