+++ code/src/sixsq/nuvla/server/resources/voucher_report.clj
@@ -22,8 +22,9 @@ direct relationship and to allow accounting on a per voucher basis.
 (def ^:const collection-type (u/ns->collection-type *ns*))
 
 
-(def collection-acl {:query ["group/nuvla-user"]
-                     :add   ["group/nuvla-user"]})
+(def collection-acl {:query       ["group/nuvla-user"]
+                     :add         ["group/nuvla-user"]
+                     :bulk-delete ["group/nuvla-user"]})
 
 
 ;;
@@ -102,3 +103,10 @@ direct relationship and to allow accounting on a per voucher basis.
   [request]
   (query-impl request))
 
+
+(def bulk-delete-impl (std-crud/bulk-delete-fn resource-type collection-acl collection-type))
+
+
+(defmethod crud/bulk-delete resource-type
+  [request]
+  (bulk-delete-impl request))