+++ code/test/sixsq/nuvla/server/resources/email_lifecycle_test.clj
@@ -199,7 +199,6 @@
 (deftest bad-methods
   (let [resource-uri (str p/service-context (u/new-resource-id t/resource-type))]
     (ltu/verify-405-status [[base-uri :options]
-                            [base-uri :delete]
                             [resource-uri :put]
                             [resource-uri :options]
                             [resource-uri :post]])))