+++ code/test/sixsq/nuvla/server/resources/evidence_record_lifecycle_test.clj
@@ -432,7 +432,6 @@
 (deftest bad-methods
   (let [resource-uri (str p/service-context (u/new-resource-id t/resource-type))]
     (ltu/verify-405-status [[base-uri :options]
-                            [base-uri :delete]
                             [resource-uri :options]
                             [resource-uri :post]
                             [resource-uri :put]])))