+++ code/test/sixsq/nuvla/server/resources/group_lifecycle_test.clj
@@ -134,6 +134,5 @@
 (deftest bad-methods
   (let [resource-uri (str p/service-context (u/new-resource-id group/resource-type))]
     (ltu/verify-405-status [[base-uri :options]
-                            [base-uri :delete]
                             [resource-uri :options]
                             [resource-uri :post]])))