+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Change
+  - Support bulk delete for more resources and fix operations for collection
+  - Job - Stale job because of race condition fix
+  - NuvlaBox - Simplify names of nuvlabox sub-resources at commission time
+
 ## [4.2.5] - 2019-10-10
 
 ### Change